2d renderer bugfixing now correctly batches & renders 2d sprites with different textures still does nothing with shaders/post processing

+++ SORE_lib/src/sore_vbo.h
@@ -29,7 +29,7 @@ namespace SORE_Graphics
 		
 		void BeginDraw();
     void DrawAll();
-		void DrawElements(unsigned int numTris, unsigned int triOffset);
+		void DrawElements(unsigned int numTris, unsigned short triOffset);
 		void EndDraw();
   private:
     GLuint vbo, vboIndices, vboNormals, vboTexCoords, vboColors;
